# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "BillingAccountCreate": {
        "additionalProperties": false,
        "properties": {
          "billing_day": {
            "maximum": 28.0,
            "minimum": 1.0,
            "title": "Billing Day",
            "type": "integer"
          },
          "currency": {
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "external_customer_ref": {
            "anyOf": [
              {
                "maxLength": 180,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Customer Ref"
          },
          "legal_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Legal Name",
            "type": "string"
          },
          "tax_profile": {
            "additionalProperties": true,
            "title": "Tax Profile",
            "type": "object"
          }
        },
        "required": [
          "legal_name",
          "currency",
          "billing_day"
        ],
        "title": "BillingAccountCreate",
        "type": "object"
      },
      "BillingAccountOut": {
        "additionalProperties": false,
        "properties": {
          "billing_day": {
            "title": "Billing Day",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "external_customer_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Customer Ref"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "legal_name": {
            "title": "Legal Name",
            "type": "string"
          },
          "owner_tenant_id": {
            "title": "Owner Tenant Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tax_profile": {
            "additionalProperties": true,
            "title": "Tax Profile",
            "type": "object"
          }
        },
        "required": [
          "id",
          "owner_tenant_id",
          "legal_name",
          "currency",
          "billing_day",
          "status",
          "tax_profile",
          "external_customer_ref",
          "created_at"
        ],
        "title": "BillingAccountOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SubscriptionOut": {
        "additionalProperties": false,
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "period_end": {
            "format": "date-time",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date-time",
            "title": "Period Start",
            "type": "string"
          },
          "plan_code": {
            "title": "Plan Code",
            "type": "string"
          },
          "quota_behavior": {
            "title": "Quota Behavior",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "suite_code": {
            "title": "Suite Code",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "trial_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial End"
          }
        },
        "required": [
          "id",
          "account_id",
          "tenant_id",
          "suite_code",
          "plan_code",
          "status",
          "quota_behavior",
          "period_start",
          "period_end",
          "trial_end",
          "created_at"
        ],
        "title": "SubscriptionOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Shared tenant, profile, identity, consent, preference, billing, usage, and entitlement authority. Effectful cross-system calls remain governed by Codestra Middleware.",
    "title": "Codestra Foundation API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/tenants/{tenant_id}/billing/accounts": {
      "post": {
        "operationId": "create_account_v1_tenants__tenant_id__billing_accounts_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 240,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": true,
            "schema": {
              "maxLength": 180,
              "minLength": 8,
              "title": "X-Correlation-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillingAccountCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingAccountOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Account",
        "tags": [
          "billing-usage"
        ]
      }
    },
    "/v1/tenants/{tenant_id}/billing/accounts/{account_id}": {
      "get": {
        "operationId": "get_account_v1_tenants__tenant_id__billing_accounts__account_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingAccountOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Account",
        "tags": [
          "billing-usage"
        ]
      }
    },
    "/v1/tenants/{tenant_id}/billing/invoices": {
      "post": {
        "operationId": "create_invoice_v1_tenants__tenant_id__billing_invoices_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 240,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": true,
            "schema": {
              "maxLength": 180,
              "minLength": 8,
              "title": "X-Correlation-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Invoice",
        "tags": [
          "billing-usage"
        ]
      }
    },
    "/v1/tenants/{tenant_id}/billing/invoices/{invoice_id}": {
      "get": {
        "operationId": "get_invoice_v1_tenants__tenant_id__billing_invoices__invoice_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "title": "Invoice Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Invoice",
        "tags": [
          "billing-usage"
        ]
      }
    },
    "/v1/tenants/{tenant_id}/billing/subscriptions": {
      "get": {
        "operationId": "list_subscriptions_v1_tenants__tenant_id__billing_subscriptions_get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubscriptionOut"
                  },
                  "title": "Response List Subscriptions V1 Tenants  Tenant Id  Billing Subscriptions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Subscriptions",
        "tags": [
          "billing-usage"
        ]
      },
      "post": {
        "operationId": "create_subscription_v1_tenants__tenant_id__billing_subscriptions_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 240,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": true,
            "schema": {
              "maxLength": 180,
              "minLength": 8,
              "title": "X-Correlation-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Subscription",
        "tags": [
          "billing-usage"
        ]
      }
    },
...
  }
}